@@ -1,6 +1,6 @@
 {
   "name": "cv",
-  "version": "1.1.34",
+  "version": "1.1.36",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -13,44 +13,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.14",
-    "@angular/cdk": "^7.3.7",
-    "@angular/cli": "^7.3.8",
-    "@angular/common": "^7.2.14",
-    "@angular/compiler": "^7.2.14",
-    "@angular/core": "^7.2.14",
-    "@angular/fire": "^5.1.3",
+    "@angular/animations": "^8.0.1",
+    "@angular/cdk": "^8.0.1",
+    "@angular/cli": "^8.0.3",
+    "@angular/common": "^8.0.1",
+    "@angular/compiler": "^8.0.1",
+    "@angular/core": "^8.0.1",
+    "@angular/fire": "^5.2.1",
     "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "^7.2.14",
-    "@angular/http": "^7.2.14",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.14",
-    "@angular/platform-browser-dynamic": "^7.2.14",
-    "@angular/router": "^7.2.14",
-    "@ngrx/effects": "^7.4.0",
-    "@ngrx/entity": "^7.4.0",
-    "@ngrx/store": "^7.4.0",
-    "@ngrx/store-devtools": "^7.4.0",
-    "core-js": "^3.0.1",
-    "firebase": "^5.11.1",
+    "@angular/forms": "^8.0.1",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "^8.0.1",
+    "@angular/platform-browser-dynamic": "^8.0.1",
+    "@angular/router": "^8.0.1",
+    "@ngrx/effects": "^8.0.1",
+    "@ngrx/entity": "^8.0.1",
+    "@ngrx/store": "^8.0.1",
+    "@ngrx/store-devtools": "^8.0.1",
+    "core-js": "^3.1.4",
+    "firebase": "^6.2.0",
     "hammerjs": "^2.0.8",
     "hoek": "^6.1.3",
     "jasmine": "^3.4.0",
     "moment": "^2.24.0",
-    "ng-recaptcha": "^4.2.1",
-    "rxjs": "^6.5.1",
-    "tslib": "^1.9.3",
-    "tslint": "^5.16.0",
-    "zone.js": "~0.8.26"
+    "ng-recaptcha": "^4.3.0",
+    "rxjs": "^6.5.2",
+    "tslib": "^1.10.0",
+    "tslint": "^5.17.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/compiler-cli": "^7.2.14",
-    "@angular/language-service": "^7.2.14",
-    "@types/jasmine": "^3.3.12",
+    "@angular-devkit/build-angular": "^0.800.3",
+    "@angular/compiler-cli": "^8.0.1",
+    "@angular/language-service": "^8.0.1",
+    "@types/jasmine": "^3.3.13",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^11.13.9",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.0.8",
+    "codelyzer": "^5.1.0",
     "jasmine-core": "^3.4.0",
     "jasmine-spec-reporter": "^4.2.1",
     "karma": "^4.1.0",
@@ -59,7 +59,7 @@
     "karma-jasmine": "^2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "^6.0.0",
-    "ts-node": "^8.1.0",
-    "typescript": "~3.2.4"
+    "ts-node": "^8.3.0",
+    "typescript": "~3.4.5"
   }
 }
